worker/requirements.txt
MetaTrader5>=5.0.45
supabase>=2.0.0
cryptography>=41.0.0
python-dotenv>=1.0.0
